+++ src/components/DataCiteWidget.vue
@@ -17,7 +17,7 @@
 
 <script>
 
-import BaseWidget from './BaseWidget.vue'
+import BaseWidget from '@/models/BaseWidget.vue'
 
 export default {
   name: 'DataCiteWidget',
