+++ src/components/MediumWidget.vue
@@ -33,7 +33,7 @@
 
 <script>
 
-import BaseWidget from './BaseWidget.vue'
+import BaseWidget from '@/models/BaseWidget.vue'
 
 export default {
   name: 'MediumWidget',
