+++ src/models/BaseWidget.vue
